class11/knowledge_graph_builder/requirements.txt

openai
langchain
graphviz
wikipedia
requests
serpapi
streamlit
langgraph
euriai
pillow
python-dotenv
pandas
fastapi
uvicorn
python-multipart
numpy
scikit-learn
